# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DHD Livraison HR Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vacation-requests:
    get:
      operationId: listVacationRequests
      tags: [vacation-requests]
      summary: List vacation requests
      parameters:
        - name: employeeId
          in: query
          schema:
            type: number
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Vacation requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VacationRequest"
    post:
      operationId: createVacationRequest
      tags: [vacation-requests]
      summary: Submit a vacation request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VacationRequestInput"
      responses:
        "201":
          description: Vacation request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VacationRequest"
  /vacation-requests/{id}/approve:
    patch:
      operationId: approveVacationRequest
      tags: [vacation-requests]
      summary: Approve vacation request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VacationRequest"
  /vacation-requests/{id}/reject:
    patch:
      operationId: rejectVacationRequest
      tags: [vacation-requests]
      summary: Reject vacation request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VacationRequest"

  # ── NOTIFICATIONS ─────────────────────────────────────────────────────────
...
  /stats/dashboard:
    get:
      operationId: getDashboardStats
      tags: [stats]
      summary: Get overall dashboard statistics
      parameters:
        - name: month
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /stats/office:
    get:
      operationId: getOfficeStats
      tags: [stats]
      summary: Get statistics for a specific office
      parameters:
        - name: officeId
          in: query
          required: true
          schema:
            type: number
        - name: month
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Office statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfficeStats"
  /stats/attendance-chart:
    get:
      operationId: getAttendanceChart
      tags: [stats]
      summary: Get attendance chart data (daily counts for a month)
      parameters:
        - name: officeId
          in: query
          schema:
            type: number
        - name: month
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChartDataPoint"
  /stats/salary-chart:
    get:
      operationId: getSalaryChart
      tags: [stats]
      summary: Get salary trends chart data (monthly totals)
      parameters:
        - name: officeId
          in: query
          schema:
            type: number
        - name: year
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Salary chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChartDataPoint"

# ── COMPONENTS ──────────────────────────────────────────────────────────────
components:
  schemas:
...
    RejectInput:
      type: object
      properties:
        reason:
          type: string

    # LEAVE REQUESTS
...
    VacationRequest:
      type: object
      required: [id, employeeId, startDate, endDate, status, requestedAt]
      properties:
        id:
          type: number
        employeeId:
          type: number
        employeeName:
          type: ["string", "null"]
        officeName:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        rejectionReason:
          type: ["string", "null"]
        requestedAt:
          type: string
          format: date-time
        resolvedAt:
          type: ["string", "null"]
          format: date-time
    VacationRequestInput:
      type: object
      required: [employeeId, startDate, endDate]
      properties:
        employeeId:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        description:
          type: string

    # VIOLATIONS
...
    DashboardStats:
      type: object
      required: [totalEmployees, presentToday, absentToday, lateToday, pendingRequests, totalSalariesThisMonth, totalAdvancesThisMonth]
      properties:
        totalEmployees:
          type: number
        presentToday:
          type: number
        absentToday:
          type: number
        lateToday:
          type: number
        pendingRequests:
          type: number
        pendingAdvances:
          type: number
        pendingLeaves:
          type: number
        pendingVacations:
          type: number
        totalSalariesThisMonth:
          type: number
        totalAdvancesThisMonth:
          type: number
        totalDeductionsThisMonth:
          type: number
        officeBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/OfficeStats"
    OfficeStats:
      type: object
      required: [officeId, officeName, totalEmployees, presentToday, absentToday, lateToday, totalSalaries, totalDeductions, totalAdvances, overtimeHours]
      properties:
        officeId:
          type: number
        officeName:
          type: string
        totalEmployees:
          type: number
        presentToday:
          type: number
        absentToday:
          type: number
        lateToday:
          type: number
        totalSalaries:
          type: number
        totalDeductions:
          type: number
        totalAdvances:
          type: number
        overtimeHours:
          type: number
    ChartDataPoint:
      type: object
      required: [label, value]
      properties:
        label:
          type: string
        value:
          type: number
        secondary:
          type: ["number", "null"]